+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/rest_api/rest_api.py
@@ -2254,7 +2254,7 @@ class DerivativesTradingUsdsFuturesRestAPI:
 
     def all_orders(
         self,
-        symbol: Union[str, None],
+        symbol: Optional[str] = None,
         order_id: Optional[int] = None,
         start_time: Optional[int] = None,
         end_time: Optional[int] = None,
@@ -2279,7 +2279,7 @@ class DerivativesTradingUsdsFuturesRestAPI:
         - The query time period must be less then 7 days( default as the recent 7 days).
 
                 Args:
-                    symbol (Union[str, None]):
+                    symbol (Optional[str] = None):
                     order_id (Optional[int] = None):
                     start_time (Optional[int] = None): Start time
                     end_time (Optional[int] = None): End time